there is not really a solution for parsing a GeoJSON object
into a Proto message. root of the problem is that you cannot
have arrays of arrays in Proto3 (and GeoJSON coordinates are
defined as arrays of arrays of arrays etc.).

So? In binary messages we will put the same bytes into the
geometry slot as we use in openLCA (question is if we should
gzip them or not... this may depends how much space we win
with gzip and how it affects the performance).

We will handle JSON-Messages a bit different => e.g. parse
them as JSON and pack binary when loading...