instead of collecting and indexing all exchanges from the
database, as it is currently done in the JSON-LD import,
we will first index the links by

process ID -> internal ID -> link

then we scan the exchnage table and assign the exchange IDs
to the links. This should be much more efficient as currently
done in the JSON-LD import (because event the elementary flows
are indexed there currently). would be interesting to measure
the time for this (also regarding the default provider update)

It is a problem when the exchange IDs are not set to values >= 0.
In this case we could still identify an exchange by the flow ID.
But this would not work for databases like exoinvent where the
same flow in a process can come from different providers (e.g.
electricity mixes). So, we can say that this is a requirement.
But then, why we do not just delete the flow ID? Is this thing
required? It is used for faster display in the product system
editor though, but not needed for data exchange... quite some
bits that we could probably save here.

testing the query:

```py

import time

from org.openlca.core.database import NativeSql

count = 0

def handle(r):
  global count
  t = (r.getLong(1), r.getLong(2), r.getInt(3))
  count += 1
  return True

for i in range(1, 11):
  count = 0
  sql = 'select id, f_owner, internal_id from tbl_exchanges'
  start = time.time()
  NativeSql.on(db).query(sql, handle)
  end = time.time()
  print('run %d: scanned %d rows in %.3f sec' % (i, count, end - start))
```

```
run 1: scanned 583207 rows in 0.960 sec
run 2: scanned 583207 rows in 0.746 sec
run 3: scanned 583207 rows in 0.715 sec
run 4: scanned 583207 rows in 0.739 sec
run 5: scanned 583207 rows in 0.760 sec
run 6: scanned 583207 rows in 0.775 sec
run 7: scanned 583207 rows in 0.737 sec
run 8: scanned 583207 rows in 0.730 sec
run 9: scanned 583207 rows in 0.723 sec
run 10: scanned 583207 rows in 0.725 sec
```