added `toFlowRef` for now
this is not so nice. but there is no inheritance in Protocol Buffers. so we
have two options: have a separate `FlowRef` type or merge everything in the
`Ref` type and only use the fields in the respective contexts... both options
are not ideal. we will keep the `FlowRef` type for now...