we could use an `In` and `Out` class
for the utility methods. `In` for incoming proto-objects, `Out` for producing
proto-objects. pushing or fetching dependencies in the imports and
exports should be more explicit and not hidden in these utility functions.
we could also have two methods like `shortRef(object)` that only contains the
necessary data and `fullRef(object, db)` for more details (this is what we need
when the client asks for providers or flow mappings)...